Trait hiring_challenge_rust::bank::accounts::AccountService

pub trait AccountService: Clone + Send + Sync + 'static {
    // Required methods
    fn place_hold<'life0, 'life1, 'async_trait>(
        &'life0 self,
        account_number: &'life1 AccountNumber,
        amount: i32
    ) -> Pin<Box<dyn Future<Output = Result<HoldRef, String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn release_hold<'life0, 'async_trait>(
        &'life0 self,
        hold_ref: HoldRef
    ) -> Pin<Box<dyn Future<Output = Result<(), String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn withdraw_funds<'life0, 'async_trait>(
        &'life0 self,
        hold_ref: HoldRef
    ) -> Pin<Box<dyn Future<Output = Result<(), String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Client to interact with a remote service that manages customer accounts.

Required Methods

fn place_hold<'life0, 'life1, 'async_trait>( &'life0 self, account_number: &'life1 AccountNumber, amount: i32 ) -> Pin<Box<dyn Future<Output = Result<HoldRef, String>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Places a hold on the account.

Reduces the account_number account’s actual balance by amount.

Placing a hold does NOT remove or transfer money from the account, it merely prevents the money from being otherwise spent until either

the money is removed from the account and sent to the proper recipient via withdraw_funds (in case the payment is concluded);

the hold is released via release_hold and the account holder may once again spec the money as they wish (in case the payment is canceled)

In other words, for every call to place_hold, there MUST be a matching call to either release_hold or withdraw_funds.

fn release_hold<'life0, 'async_trait>( &'life0 self, hold_ref: HoldRef ) -> Pin<Box<dyn Future<Output = Result<(), String>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Releases a hold on the account.

Increases the account_number account’s actual balance by the amount previously held.

Typically, this is used when the payment for which the hold was created doesn’t go through fully (it was canceled, the system failed, etc.). Unless holds are released, a failed payment would mean that the customer wouldn’t get the goods (because the merchant wasn’t paid), but wouldn’t have access to his money either because a hold is still present on the funds.

fn withdraw_funds<'life0, 'async_trait>( &'life0 self, hold_ref: HoldRef ) -> Pin<Box<dyn Future<Output = Result<(), String>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Withdraws the held money from the account.

Decreases the current balance of the account linked to the hold reference by the amount previously held. The hold on the customer’s funds is implicitly released atomically.

This is the mechanism by which money is transferred out from the customer’s account and into the merchant’s account during the settlement process.